+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
-- _Nothing yet._
+### Fixed
+
+- Parts menu array submenus are no longer clipped to the tracker height.
 
 ## [1.1.2] - 2026-01-28